OEM: Provide Proxmox images
This uses https://github.com/coreos/afterburn/pull/1023 and
https://github.com/coreos/ignition/pull/1790 to provide
Proxmox images.

This pulls in https://github.com/flatcar/bootengine/pull/91
and https://github.com/flatcar/init/pull/115 to run afterburn for
hostname, network, SSH key, and metadata attribute setup. The afterburn
support for the SSH key and hostname parses the user-data when it's
cloud-init. The coreos-cloudinit support is not there but can be added
in addition: We need to add a new provider that varies from the existing
config drive support because the file is called user-data and not
user_data, and it needs to look for a filesystem label cidata and not
config-2.